NotFoundException

+++ AdessoRideShare.Application/Trips/Commands/DeleteTrip/DeleteTripCommandHandler.cs
@@ -1,4 +1,6 @@
-﻿using AdessoRideShare.Application.Common.Interfaces;
+﻿using AdessoRideShare.Application.Common.Exceptions;
+using AdessoRideShare.Application.Common.Interfaces;
+using AdessoRideShare.Domain.Entities;
 using MediatR;
 using Microsoft.EntityFrameworkCore;
 using System.Threading;
@@ -18,6 +20,11 @@ namespace AdessoRideShare.Application.Trips.Commands.DeleteTrip
         public async Task<Unit> Handle(DeleteTripCommand request, CancellationToken cancellationToken)
         {
             var trip = await context.Trips.SingleOrDefaultAsync(x => x.Id == request.Id);
+
+            if (trip == null)
+            {
+                throw new NotFoundException(nameof(Trip), request.Id);
+            }
 
             context.Trips.Remove(trip);
             await context.SaveChangesAsync(cancellationToken);
